refactor(actions): consolidate profileEntriesEqual into a shared module (PR review)
